services/functions/slo-rollback-fn/requirements.txt
# This function uses the google-cloud-firestore library to
# manage the debounce timestamp.
google-cloud-firestore
